% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SCE.R
\name{association.SingleCellExperiment}
\alias{association.SingleCellExperiment}
\title{Extract information from Seurat object to create CNA object}
\usage{
association.SingleCellExperiment(
  sce,
  test_var,
  sample_key,
  graph,
  batches = NULL,
  covs = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sce}{Initialized SingleCellExperiment object.}

\item{test_var}{Variable to test for association.}

\item{sample_key}{String denoting the name of the sample-level identifier (e.g. DonorID).}

\item{graph}{A KNN graph in sparseMatrix format or the name of such a graph
stored in `colPairs(sce)`.}

\item{batches}{A character string. Name of batch variable. Currently only one 
categorical variable allowed.}

\item{covs}{A character string or vector. Name(s) of other (numerical) covariates 
to control for.}

\item{verbose}{Logical controlling verbosity}

\item{...}{Passed to [association()]}
}
\value{
A seurat object
}
\description{
Extract information from Seurat object to create CNA object
}
