% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{batchKurtosis}
\alias{batchKurtosis}
\title{Calculate kurtosis accounting for batch}
\usage{
batchKurtosis(NAM, batch)
}
\arguments{
\item{NAM}{An M x N (Cells x Samples) neighborhood association matrix as 
calculated by [buildNAM()]}

\item{batch}{A factor containing batch information}
}
\value{
a vector of kurtosis values for each column (cell) in the NAM
}
\description{
Calculate kurtosis accounting for batch
}
\keyword{internal}
