% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Seurat.R
\name{create_object.Seurat}
\alias{create_object.Seurat}
\title{Extract information from Seurat object to create CNA object}
\usage{
create_object.Seurat(
  seurat_object,
  samplem_key,
  samplem_vars,
  graph_use = NULL
)
}
\arguments{
\item{seurat_object}{Initialized Seurat object. Assumes that FindNeighborhoods has been run.}

\item{samplem_key}{String denoting the name of the sample-level identifier (e.g. DonorID).}

\item{samplem_vars}{Which sample-level covariates to include.}

\item{graph_use}{Which graph to use. By default, will use first graph in seurat_object.}
}
\value{

}
\description{
Extract information from Seurat object to create CNA object
}
