% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createCNAListSCE}
\alias{createCNAListSCE}
\title{Extract information from SingleCellExperiment object to create a CNA object}
\usage{
createCNAListSCE(sce, y, sample.key, sample.vars = NULL, knn.graph)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{y}{A character string denoting the name of the variable to be 
tested for association.}

\item{sample.key}{A character string denoting the name of the sample-level 
identifier (e.g. DonorID).}

\item{sample.vars}{A character string or vector specifying which columns of 
`colData(sce)` should be included. This should include any columns for batch
and covariates.}

\item{knn.graph}{A KNN graph in `sparseMatrix` format, or the name of such a graph
stored in `colPair(sce)`.}
}
\value{
An `rnca` data object (list)
}
\description{
Extract information from SingleCellExperiment object to create a CNA object
}
