% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SCE.R
\name{createObject.SingleCellExperiment}
\alias{createObject.SingleCellExperiment}
\title{Extract information from SingleCellExperiment object to create CNA object}
\usage{
createObject.SingleCellExperiment(sce, sample_key, sample_vars, graph)
}
\arguments{
\item{sce}{Initialized Seurat object. Assumes that FindNeighborhoods has been run.}

\item{sample_key}{Character string denoting the name of the sample-level identifier (e.g. DonorID).}

\item{sample_vars}{Character string or vector specifying Which columns of 
`colData(sce)` should be included. This should include any columns for batch
and covariates.}

\item{graph}{A KNN graph in sparseMatrix format or the name of such a graph
stored in `colPairs(sce)`.}
}
\value{
an rnca data object (list)
}
\description{
Extract information from SingleCellExperiment object to create CNA object
}
