% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{diffuseStep}
\alias{diffuseStep}
\title{Perform a single diffusion mapping step.}
\usage{
diffuseStep(a, s)
}
\arguments{
\item{a}{A sparseMatrix M by M (cells x cells) adjacency matrix (connectivities) 
whose m,m'-th entry indicates the similarity between cells m and m' in the graph}

\item{s}{A sparseMatrix M by N (cells x samples) indicator matrix assigning 
the m-th cell to the n-th sample from which it was drawn.}
}
\value{
A M by N sparseMatrix.
}
\description{
Perform a single diffusion mapping step.
}
\note{
This function ignores distances and only uses unweighted connectivities
}
\keyword{internal}
