% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dimplot.generic}
\alias{dimplot.generic}
\title{Scatterplot with color}
\usage{
dimplot.generic(
  coords,
  y,
  label = NULL,
  color.low = "blue",
  color.high = "red"
)
}
\arguments{
\item{coords}{Coordinate table}

\item{y}{Vector for colors}

\item{label}{(optional) name for y vector.\cr
Default: NULL}

\item{color.low, color.high}{Low and high color for correlation values.\cr
Default: muted("blue") and muted("red"), respectively}
}
\value{
A ggplot object
}
\description{
Scatterplot with color
}
