% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dimplot.ncorr}
\alias{dimplot.ncorr}
\title{Scatter plot of correlation values based on FDR threshold}
\usage{
dimplot.ncorr(
  coords,
  res,
  fdr_thresh = 0.05,
  color.low = "blue",
  color.high = "red"
)
}
\arguments{
\item{coords}{A data.frame of coordinates (x,y).}

\item{res}{A result output from [association] function}

\item{fdr_thresh}{A numeric FDR threshold (\eqn{\alpha}), values larger than this
are colored gray.}

\item{color.low, color.high}{Low and high color for correlation values.\cr
Default: muted("blue") and muted("red"), respectively}
}
\value{
A ggplot object
}
\description{
Scatter plot of correlation values based on FDR threshold
}
