% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dimplotCoords}
\alias{dimplotCoords}
\title{Scatterplot with color}
\usage{
dimplotCoords(
  coords,
  y,
  label = NULL,
  reorder = c("shuffle", "signif"),
  threshold = NULL,
  color.low = "blue3",
  color.high = "red3",
  point.size = 0.5,
  point.alpha = 0.5
)
}
\arguments{
\item{coords}{A matrix or dataframe with at least 2 columns containing the 
x and y coordinates to plot.}

\item{y}{A vector of values to use for coloring each point in coords.}

\item{label}{(optional) A character string to name the y vector.\cr
Default: NULL}

\item{reorder}{A character string specifying how the points should be ordered.
One of "shuffle" or "signif" where "shuffle" randomly shuffles the points to 
produce an even distribution of values and "signif" puts the largest values of
y on top.}

\item{threshold}{An (optional) numerical value below which values of y are 
not colored. Note that abs(y) > threshold is checked rather than y > threshold. \cr
Default: NULL}

\item{color.low, color.high}{Low and high color for correlation values.\cr
Default: "blue4" and "red4", respectively}

\item{point.size, point.alpha}{Numeric values controlling point size and alpha.\cr
Defaults: 0.5 (both)}
}
\value{
A ggplot object
}
\description{
Generic 2D plot with coloring. Takes a set of coordinates and y, where y is used
to color each row in coords.
}
