% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dimplotCorr}
\alias{dimplotCorr}
\title{Scatter plot of correlation values based on FDR threshold}
\usage{
dimplotCorr(
  sce,
  res,
  dim.use = "CNA",
  label = "corr",
  fdr.thresh = 0.05,
  reorder = "signif",
  color.low = "blue3",
  color.high = "red3",
  point.size = 0.5,
  point.alpha = 0.5
)
}
\arguments{
\item{sce}{An sce object to which has the results of [association] via 
[associationSCE]. Mutually exclusive with `res`.}

\item{res}{A result output from [association] function. Mutually exclusive 
with `sce`.}

\item{dim.use}{A character string or NULL. \cr
Default: "CNA"}

\item{label}{(optional) A character string to name the y vector, and by 
extension, the legend.\cr
Default: "corr"}

\item{fdr.thresh}{A numeric FDR threshold (\eqn{\alpha}), values larger than 
this are colored gray.}

\item{reorder}{A characterstring specifying whether the data should be reordered
to place significant values on top ("signif") or randomly shuffle the data ("shuffle").\cr
Default: "signif"}

\item{color.low, color.high}{Low and high color for correlation values.\cr
Default: "blue3" and "red3", respectively}

\item{point.size, point.alpha}{Numeric values controlling point size and alpha.\cr
Defaults: 0.5 (both)}
}
\value{
A ggplot object
}
\description{
Scatter plot of correlation values based on FDR threshold
}
