% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{innerNAM}
\alias{innerNAM}
\title{Calculate Neighborhood Association Matrix (NAM) using diffusion mapping}
\usage{
innerNAM(data, n.steps = NULL, min.steps = 3L, max.steps = 15L, kurt.delta = 3)
}
\arguments{
\item{data}{A cna data list.}

\item{n.steps}{Number of steps to take during the random walk.\cr
Default: NULL}

\item{min.steps}{minimum number of steps to take if n.steps is NULL.\cr
Default: 3}

\item{max.steps}{Maximum number of steps to take if n.steps is NULL.\cr
Default: 15}

\item{kurt.delta}{the minimum change}
}
\value{
A sparseMatrix containing the NAM
}
\description{
Calculate Neighborhood Association Matrix (NAM) using diffusion mapping
}
\note{
The original code had a bug here where `prevmedkurt <- medkurt` was 
updated before calculating `prevmedkurt - medkurt` and so only a single diffusion
step was performed.
}
\keyword{internal}
