% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{olsNAM}
\alias{olsNAM}
\title{Perform batch correction of a NAM with OLS regression}
\usage{
olsNAM(NAM, cov.mat = NULL, batch = NULL)
}
\arguments{
\item{NAM}{A neighborhood association matrix}

\item{cov.mat}{a matrix of covariates. \cr
Default: NULL}

\item{batch}{a factor containing the batch to correct. \cr
Default: NULL}
}
\value{
A list containing the corrected NAM, the Hat matrix (M), and degrees 
of freedom (r)
}
\description{
Perform batch correction of a NAM with OLS regression
}
\keyword{internal}
