% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sm.R
\name{propTable}
\alias{propTable}
\title{A prop.table alternative for sparseMatrices}
\usage{
propTable(x, margin)
}
\arguments{
\item{x}{A sparseMatrix}

\item{margin}{A vector giving the margins to split by. E.g., for a matrix 1 
indicates rows, 2 indicates columns.}
}
\value{
A proportion table for x over the given margin in sparseMatrix format.
}
\description{
A prop.table alternative for sparseMatrices
}
\note{
Unlike [base::prop.table()], joint normalization (`margin = c(1, 2)`)
is not supported. Rows or columns that sum to zero will produce `NaN`.
}
\keyword{internal}
