% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{RCNA_analysis-class}
\alias{RCNA_analysis-class}
\alias{RCNA_analysis}
\title{RCNA_analysis constructor}
\description{
An S4 class used to track parameters from a specific RCNA function execution.
}
\section{Slots}{

\describe{
\item{\code{call}}{A character vector detailing the function call that (`correct_gc_bias`, `estimate_nkr`, `estimate_feature_score`) was performed to produce this S4 object.}

\item{\code{params}}{A list corresponding to the parameters that were submitted with the associated function call}

\item{\code{res.files}}{A list containing the names of the flat files that were created by the documented function call.}
}}

