% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcna_demo}
\alias{rcna_demo}
\title{Demo data}
\format{
:
  connectivities: dgCMatrix of knn indices (1e4x1e4)
  samplem: sample-level meta data (n=50 rows)
  obs: cell-level meta data (n=1e4 rows)
  obsm: list of misc metadata. Contains umap coords table. 
  N: number of cells (=1e4)
}
\source{
\url{github.com/yakirr/cna/tree/master/demo}
}
\usage{
rcna_demo
}
\description{
Demo data
}
\keyword{datasets}
