% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\docType{class}
\name{RCNA_object-class}
\alias{RCNA_object-class}
\alias{RCNA_object}
\title{RCNA.object definition}
\description{
An S4 class used to specify parameters for an analysis run
}
\section{Slots}{

\describe{
\item{\code{sample.names}}{Required. Character vector containing names of subjects}

\item{\code{ano.file}}{Required. Character single file path detailing a feature-wise annotation file}

\item{\code{out.dir}}{Required. Character vector containing the name of each subject?s output directory}

\item{\code{gcParams}}{Data Frame storing all run parameters for the correct_gc_bias function}

\item{\code{nkrParams}}{Data Frame storing all run parameters for the estimate_nkr function}

\item{\code{scoreParams}}{Data Frame storing all run parameters for the estimate_feature_score function}

\item{\code{commands}}{RCNA_analysis object storing commands and parameters from previous function runs on this object}
}}

\seealso{
\code{\link{run_RCNA}}
\code{\linkS4class{RCNA_analysis}},
}
