% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{ridgeNAM}
\alias{ridgeNAM}
\title{Perform batch correction of a NAM using (partial) ridge regression}
\usage{
ridgeNAM(
  NAM,
  cov.mat = NULL,
  batch = NULL,
  ridges = NULL,
  ridge.crit = c("kurtosis", "gcv"),
  partial = NULL,
  verbose = FALSE
)
}
\arguments{
\item{NAM}{An M x N (Cells x Samples) neighborhood association matrix as 
calculated by [buildNAM()]}

\item{cov.mat}{a matrix of covariates. \cr
Default: NULL}

\item{batch}{a factor containing the batch to correct. \cr
Default: NULL}

\item{ridges}{A vector of ridges to try or NULL. If NULL ridges are calculated
automatically centered on `svd(X)$d[1]^2`. \cr
Default: NULL}

\item{ridge.crit}{A character string specifying how the ridge regression should
select optimal ridges. One of c("gcv", "kurtosis"). Use "kurtosis" to match
original cna implementation.
# Default: "kurtosis"}

\item{partial}{A character string or NULL. If NULL, ridge regression is performed
on all variables. If a character, specify which set of variables ridge regression
should be used on, the other set does not get penalized. Must be one of "batches" or
"covariates". \cr 
Default: NULL}

\item{verbose}{Logical controlling verbosity.\cr
Default: FALSE}
}
\value{
A list containing the corrected NAM, the Hat matrix (M), and the 
effective degree of freedom (`sum(diag(H))` from the ridge hat matrix) (r)
}
\description{
Perform batch correction of a NAM using (partial) ridge regression
}
\keyword{internal}
