% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{svdNAM}
\alias{svdNAM}
\title{Compute the singular-value decomposition of a NAM matrix.}
\usage{
svdNAM(NAM, n.pcs = NULL)
}
\arguments{
\item{NAM}{An M x N (Cells x Samples) neighborhood association matrix as 
calculated by [buildNAM()]}

\item{n.pcs}{maximum number of components to use in the SVD. Passed to the `k`
parameter of [RSpectra::svds()]. If NULL or greater than `0.5 * min(dim(NAM))`
then the full svd is calculated. \cr
Default: NULL}
}
\value{
A list containing the U, D, and V' components of the SVD.
}
\description{
Compute the singular-value decomposition of a NAM matrix.
}
\note{
svs are actually eigenvalues, not SVs. I squared them to be consistent 
with python code.

svdNAM scales NAM columns, so this is correlation, not covariance
}
\keyword{internal}
