% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tail_counts}
\alias{tail_counts}
\title{calculate the number of correlations within each threshold}
\usage{
tail_counts(breaks, z)
}
\arguments{
\item{breaks}{A vector of thresholds}

\item{z}{A matrix of correlations}
}
\value{
A numeric matrix of dimensions (\code{length(breaks) - 1} x \code{ncol(z)}), 
where entry [i, j] is the number of rows in column j of \code{z} whose squared 
value exceeds \code{breaks[i]^2}.
}
\description{
calculate the number of correlations within each threshold
}
\keyword{internal}
