% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_association.R
\name{association}
\alias{association}
\title{Main function to perform CNA association}
\usage{
association(
  data,
  y,
  batches = NULL,
  covs = NULL,
  nsteps = NULL,
  suffix = "",
  force_recompute = FALSE,
  return_nam = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{list containing samplem (sample-level metadata), 
obs (cell-level metadata), and connectivities (dgCMatrix)}

\item{y}{vector with contrast variable value for association}

\item{batches}{string(s) to denote batch variables.}

\item{covs}{string(s) to denote covariate variables.}

\item{nsteps}{TBD}

\item{suffix}{TBD}

\item{force_recompute}{TBD}

\item{verbose}{TBD}
}
\value{
TBD
}
\description{
Main function to perform CNA association
}
