% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SCE.R
\name{association.SingleCellExperiment}
\alias{association.SingleCellExperiment}
\title{Perform CNA on and assign the results to the input SingleCellExperiment object.}
\usage{
association.SingleCellExperiment(
  sce,
  test_var,
  sample_key,
  graph,
  batches = NULL,
  covs = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{test_var}{Variable to test for association.}

\item{sample_key}{String denoting the name of the sample-level identifier (e.g. DonorID).}

\item{graph}{A KNN graph in sparseMatrix format or the name of such a graph
stored in `colPair(sce)`.}

\item{batches}{A character string. Name of batch variable. Currently only one 
categorical variable allowed.}

\item{covs}{A character string or vector. Name(s) of other (numerical) covariates 
to control for.}

\item{verbose}{Logical controlling verbosity}

\item{...}{Passed to [association()]}
}
\value{
A SingleCellExperiment object with NAM embeddings stored in 
`reducedDim(sce, "CNA")`, results stored in `metadata(sce)$CNA`, and correlation
values stored in `colData(sce)` as `cna_ncorrs`, `cna_ncorrs_fdr05`, and 
`cna_ncorrs_fdr10`
}
\description{
Perform CNA on and assign the results to the input SingleCellExperiment object.
}
