% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{batchKurtosis}
\alias{batchKurtosis}
\title{calculate kurtosis accounting for batch}
\usage{
batchKurtosis(NAM, batch)
}
\arguments{
\item{NAM}{A NAM matrix as calculated by [buildNAM()]}

\item{batch}{A factor containing batch information}
}
\value{
a vector of kurtosis values for each row (cell) in NAM
}
\description{
calculate kurtosis accounting for batch
}
\keyword{internal}
