% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{buildNAM}
\alias{buildNAM}
\title{Calculate Neighborhood Association Matrix (NAM) using diffusion mapping}
\usage{
buildNAM(
  data,
  n.steps = NULL,
  min.steps = 3L,
  max.steps = 15L,
  kurt.delta = 3,
  verbose = FALSE
)
}
\arguments{
\item{data}{A rcna data object (list). See [nam()] or [association()] for details.}

\item{n.steps}{A numeric, the number of steps to take during the random walk. 
If NULL, then `kurt.delta` is checked at each step to decide whether or not 
to take the next step. \cr
Default: NULL}

\item{min.steps}{A numeric value controlling the minimum number of steps to take. 
Only used if n.steps is NULL. \cr
Default: 3}

\item{max.steps}{A numeric value controlling the maximum number of steps to take. 
Only used if n.steps is NULL. \cr
Default: 15}

\item{kurt.delta}{The minimum change in kurtosis between steps under which
the walk stops (if min.steps have already been made). Only used if n.steps is NULL.\cr
Default: 3}
}
\value{
A sparseMatrix containing the NAM (A cell by sample neighborhood
association matrix).
}
\description{
Calculate Neighborhood Association Matrix (NAM) using diffusion mapping
}
\note{
The original code had a bug here where `prevmedkurt <- medkurt` was 
updated before calculating `prevmedkurt - medkurt` and so only a single diffusion
step was performed.
}
\keyword{internal}
