% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_association.R
\name{calcStats}
\alias{calcStats}
\title{Calculate the F-test p-value}
\usage{
calcStats(yhat, ycond, k, n, r, log.p = FALSE)
}
\arguments{
\item{yhat}{y fit to the data}

\item{ycond}{y corrected for batch and covariates}

\item{k}{Number of components of U to test}

\item{n}{number of samples}

\item{r}{degrees of freedom}

\item{log.p}{A logical whether the log transformed p-value value should be 
calculated.\cr
Default: FALSE}
}
\value{
A list containing p-value and r^2 for the F-test
}
\description{
Calculate the F-test p-value
}
\keyword{internal}
