% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{conditional_permutation}
\alias{conditional_permutation}
\title{Permute Y conditioned on B}
\usage{
conditional_permutation(B, Y, num, duplicates.ok = TRUE)
}
\arguments{
\item{B}{a numeric batch vector}

\item{Y}{the variable of interest, scaled}

\item{num}{Number of permutations to perform}
}
\value{
A matrix of permutations of Y conditioned on B
}
\description{
Permute Y conditioned on B
}
\keyword{internal}
