% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_SCE.R
\name{createObject.SingleCellExperiment}
\alias{createObject.SingleCellExperiment}
\title{Extract information from SingleCellExperiment object to create CNA object}
\usage{
createObject.SingleCellExperiment(sce, sample_key, sample_vars, graph)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{sample_key}{A character string denoting the name of the sample-level 
identifier (e.g. DonorID).}

\item{sample_vars}{A character string or vector specifying which columns of 
`colData(sce)` should be included. This should include any columns for batch
and covariates.}

\item{graph}{A KNN graph in sparseMatrix format, or the name of such a graph
stored in `colPair(sce)`.}
}
\value{
An `rnca` data object (list)
}
\description{
Extract information from SingleCellExperiment object to create CNA object
}
