% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{dimplot.generic}
\alias{dimplot.generic}
\title{Scatterplot with color}
\usage{
dimplot.generic(coords, y, label = NULL)
}
\arguments{
\item{coords}{Coordinate table}

\item{y}{Vector for colors}

\item{label}{(optional) name for y vector}
}
\value{
ggplot object
}
\description{
Scatterplot with color
}
