% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{dimplot.ncorr}
\alias{dimplot.ncorr}
\title{Scatter plot of correlation values based on FDR threshold}
\usage{
dimplot.ncorr(coords, res, fdr_thresh = 0.05)
}
\arguments{
\item{coords}{A data.frame of coordinates (x,y).}

\item{res}{A result output from [association] function}

\item{fdr_thresh}{A numeric FDR threshold (\alpha), values larger than this
are colored gray.}
}
\value{
A ggplot object
}
\description{
Scatter plot of correlation values based on FDR threshold
}
