% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{dimplot.ncorr}
\alias{dimplot.ncorr}
\title{Scatterplot with ncorr and FDR}
\usage{
dimplot.ncorr(coords, res, fdr_thresh = 0.05)
}
\arguments{
\item{coords}{Coordinate table}

\item{res}{result from association function}

\item{fdr_thresh}{FDR threshold}
}
\value{
ggplot object
}
\description{
Scatterplot with ncorr and FDR
}
