% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{empirical_fdrs}
\alias{empirical_fdrs}
\title{compute the empirical FDRs via FDP procedure}
\usage{
empirical_fdrs(z, znull, thresholds)
}
\arguments{
\item{z}{A correlation matrix with a single column containing the observed 
correlations}

\item{znull}{A correlation matrix with n-permutations columns containing the 
permuted null correlations}

\item{threshold}{a vector of bins for the distribution}
}
\value{
A data.frame containing thresholds, FDRs, and number of neighborhoods in each
threshold bin.
}
\description{
compute the empirical FDRs via FDP procedure
}
\keyword{internal}
