% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_association.R
\name{innerAssociation}
\alias{innerAssociation}
\title{Calculate global and (optionally) local association tests.}
\usage{
innerAssociation(
  NAMsvd,
  y,
  batches_vec,
  ks = NULL,
  Nnull = 1000,
  force_permute_all = FALSE,
  allow_duplicate_perms = TRUE,
  local_test = TRUE,
  seed = NULL
)
}
\arguments{
\item{NAMsvd}{The list output from the [nam()] function.}

\item{y}{A vector with contrast variable value to be tested for association.}

\item{batches_vec}{A factor or numeric vector of batches to adjust for.}

\item{ks}{A numeric scalar selecting the number of components of the SVD
to test for global association. If null 4 values for k are selected between
n/50 and n/5. \cr
Default: NULL}

\item{Nnull}{A numeric specifying the number of null permutations.\cr
Default: 1000}

\item{force_permute_all}{A logical controlling whether permutations of `y` 
should preserve batch information. \cr
Default: FALSE}

\item{local_test}{A logical, whether or not to perform local test of neighborhood
correlations. \cr
Default: TRUE}

\item{seed}{A numeric seed to set. Set if you want to repeat exact permutations
as a previous run. If NULL, a random seed is chosen. \cr
Default: NULL}
}
\value{
A list. 
list(p = pfinal, 
minps_null = minps_null,
k = k,
ncorrs = ncorrs, 
fdrs = fdrs,
fdr_5p_t = fdr_5p_t, 
fdr_10p_t = fdr_10p_t,
yhat = yhat, 
ycond = ycond,
ks = ks, 
beta = beta,
r2 = r2, 
r2_perpc = r2_perpc,
nullr2_mean = mean(nullr2s), 
nullr2_std = sd(nullr2s))
}
\description{
Calculate global and (optionally) local association tests.
}
\keyword{internal}
