% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{Kurtosis}
\alias{Kurtosis}
\title{calculate kurtosis of the rows or columns of a sparseMatrix}
\usage{
Kurtosis(x, margin)
}
\arguments{
\item{x}{A sparseMatrix for which the row or column kurtosis should be calculated}

\item{margin}{An integer specifying the margin over which to calculate kurtosis
1 for rows and 2 for columns.}
}
\value{
the row- or column-wise kurtosis of x
}
\description{
calculate kurtosis of the rows or columns of a sparseMatrix
}
\keyword{internal}
