% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{minpStats}
\alias{minpStats}
\title{Calculate the p-value for all ks and return the minimum}
\usage{
minpStats(M, y, ks, U, r)
}
\arguments{
\item{M}{The annihilator matrix (M returned by [residNAM])}

\item{y}{A scaled numeric vector containing the variable of interest}

\item{ks}{A vector of ks to test.}

\item{U}{a matrix whose columns contain the left singular vectors of x 
(U returned by [svdNAM])}

\item{r}{the degrees of freedom already consumed by batch correction 
(r returned by [residNAM])
Default: FALSE}
}
\value{
A list containing the k, p, and r^2 for k which produces the minimum
p-value.
}
\description{
Calculate the p-value for all ks and return the minimum
}
\keyword{internal}
