% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{nam}
\alias{nam}
\title{Build and decompose a neighborhood abundance matrix}
\usage{
nam(
  data,
  y,
  batches = NULL,
  covs = NULL,
  n.steps = NULL,
  min.steps = 3L,
  max.steps = 15L,
  kurtosis.delta = 3,
  min.batch.kurtosis = 6,
  max.frac.pcs = 0.15,
  method = c("ridge", "ols"),
  partial.by = NULL,
  ridges = NULL,
  suffix = "",
  filter.samples = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{a list containing: 
samplem = (sample-level metadata), 
obs = (cell-level metadata), 
connectivities  = (sparse symmetric weighted adjacency matrix),
samplem_key = (character string indicating the column in samplem uniquely identifying samples),
obs_key = (character string indicating the column in obs uniquely identifying cells),
N = nrow(samplem_df).}

\item{y}{A character string specifying the column in `data$samplem` where 
the variable of interest is stored.\cr}

\item{batches}{A character string denoting the column in `data$samplem` where
batch information is stored. Only a single batch variable is allowed. \cr
Default: NULL}

\item{covs}{A character string or vector denoting the column(s) `data$samplem`
where continuous covariate information is stored.\cr
Default: NULL}

\item{n.steps}{Numeric scalar controlling the number of steps to take during 
the random walk.\cr
Default: NULL}

\item{min.steps}{Numeric scalar. Minimum number of steps to take on the random
walk.\cr
Default: 3L}

\item{max.steps}{Numeric scalar. Maximum number of steps to take on the random
walk.\cr
Default: 15L}

\item{kurtosis.delta}{Numeric scalar. Minimum kurtosis change below which the
random walk can stop (after travelling at least `min.steps`).\cr
Default: 3}

\item{min.batch.kurtosis}{Sets the minimum threshold below which cells are kept
during QC. This ensures that when kurtosis is not extreme between batches
then cells are not filtered. `threshold <- max(min.batch.kurtosis, 2*median(kurtoses))`.\cr
Default: 6}

\item{max.frac.pcs}{The number of PCs to calculate for SVD. The minimum is 10 
and the maximum is the number of samples minus 1. \cr
Default: 0.15}

\item{method}{A character string specifying what method to use for batch
and covariate correction. Options are: c("ridge", "ols"), Setting to "ridge"
will reproduce original behavior.}

\item{partial.by}{A character string. One of NULL, "batches", or "covariates". If 
`method == "ridge"` then ridge penalties are only applied to those variables.
If null then all variables are penalized. Setting to "batches" will reproduce
original behavior.\cr
Default: NULL}

\item{ridges}{A numeric vector of ridges to try during ridge regression. When
NULL ridges are calculated automatically. \cr
Default: NULL}

\item{suffix}{A character scalar for optionally adding a suffix to the output
components. \cr
Default: ""}

\item{filter.samples}{STUB. Not used currently.}

\item{verbose}{Logical scalar controlling verbosity. \cr
Default: FALSE}

\item{...}{Additional parameters passed to [buildNAM()].}
}
\value{
A list containing the NAM and associated variables.
}
\description{
Build and decompose a neighborhood abundance matrix
}
