% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{nam}
\alias{nam}
\title{Build and decompose neighborhood abundance matrix}
\usage{
nam(
  data,
  batches = NULL,
  covs = NULL,
  filter_samples = NULL,
  nsteps = NULL,
  max_frac_pcs = 0.15,
  suffix = "",
  force_recompute = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{list containing samplem (sample-level metadata), 
obs (cell-level metadata), and connectivities (dgCMatrix)}

\item{batches}{string(s) to denote batch variables.}

\item{covs}{string(s) to denote covariate variables.}

\item{filter_samples}{TBD}

\item{nsteps}{TBD}

\item{max_frac_pcs}{TBD}

\item{suffix}{TBD}

\item{force_recompute}{TBD}

\item{verbose}{TBD}
}
\value{
TBD
}
\description{
Build and decompose neighborhood abundance matrix
}
