% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sm.R
\name{propTable}
\alias{propTable}
\title{A prop.table alternative for sparseMatrices}
\usage{
propTable(x, margin)
}
\arguments{
\item{x}{A sparseMatrix}

\item{margin}{A vector giving the margins to split by. E.g., for a matrix 1 
indicates rows, 2 indicates columns.}
}
\value{
A sparse proportion table for the given margin
}
\description{
A prop.table alternative for sparseMatrices
}
\note{
unlike [base::prop.table()] this function doesn't handle `marge = c(1,2)`
}
\keyword{internal}
