% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nam.R
\name{qcNAM}
\alias{qcNAM}
\title{Filter NAM based on batch level kurtosis}
\usage{
qcNAM(NAM, batch = NULL, donors = NULL, min.threshold = 6, verbose = FALSE)
}
\arguments{
\item{NAM}{An M x N (Cells x Samples) neighborhood association matrix as 
calculated by [buildNAM()]}

\item{batch}{A factor containing batch information or NULL. If NULL, no
batch QC is performed (this is for pass-through). \cr
Default: NULL}

\item{min.threshold}{A numeric scalar defining the minimum value which the 
threshold should take.\cr
Default: 6}

\item{verbose}{Logical. Controls verbosity.\cr
Default: FALSE}
}
\value{
A list containing the filtered NAM (NAM) and an index vector indicating
which cells were kept (keep).
}
\description{
Filter NAM based on batch level kurtosis
}
\note{
min.threshold is 6 here, to keep functionality the same as initially 
written, but there is no mention of this in the paper.
}
\keyword{internal}
