% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{qcNAM}
\alias{qcNAM}
\title{Filter NAM based on batch level kurtosis}
\usage{
qcNAM(NAM, batch = NULL, min.threshold = 6, verbose = FALSE)
}
\arguments{
\item{NAM}{a sparseMatrix NAM generated by [buildNAM()]}

\item{batch}{A factor containing batch information}

\item{min.threshold}{A numeric scalar defining the minimum value which the 
threshold should take.\cr
Default: 6}

\item{verbose}{Logical. Controls verbosity.}
}
\value{
A list containing the filtered NAM (NAM) and a logical vector indicating
which cells were kept (keep).
}
\description{
Filter NAM based on batch level kurtosis
}
\note{
min.threshold is 6 here, to keep functionality the same as initially 
written, but there is no mention of this in the paper.
}
\keyword{internal}
