% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_association.R
\name{regress}
\alias{regress}
\title{Fit y to the data}
\usage{
regress(q, k, U)
}
\arguments{
\item{q}{A vector of values to fit. A column-scaled numeric vector of length n (samples)}

\item{k}{Number of components of U to fit, `1 <= k <= ncol(U)`.}

\item{U}{the left singular vector matrix of SVD (samples × PCs).}
}
\value{
a list containing fitted values (qhat) and regression coefficients (beta)
}
\description{
Fit y to the data
}
\keyword{internal}
