% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp2.r, R/utils_association.R
\name{regress}
\alias{regress}
\title{Fit y to the data}
\usage{
regress(q, k, U)

regress(q, k, U)
}
\arguments{
\item{q}{A vector of values to fit (a la y)}

\item{k}{Number of components of U to fit}

\item{U}{the left singular vector matrix of SVD}
}
\value{
a list containing fitted values (qhat) and regression coefficients (beta)

a list containing fitted values (qhat) and regression coefficients (beta)
}
\description{
Fit y to the data

Fit y to the data
}
\keyword{internal}
