% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{residNAM}
\alias{residNAM}
\title{remove batch effects (including continuous covariates)}
\usage{
residNAM(
  NAM,
  covs.mat = NULL,
  batch = NULL,
  method = c("ridge", "ols"),
  partial.by = c("batches", "covariates"),
  ridges = NULL,
  verbose = FALSE
)
}
\arguments{
\item{NAM}{An M x N (Cells x Samples) neighborhood association matrix as 
calculated by [buildNAM()]}

\item{covs.mat}{A matrix of covariates.\cr
Default NULL}

\item{batch}{A factor containing batch information.\cr
Default NULL}

\item{method}{A character string specifying what method to use for batch
and covariate correction. Options are: c("ridge", "ols"), Setting to "ridge"
will reproduce original `rcna` behavior.}

\item{partial.by}{A character string. One of NULL, "batches", or "covariates". 
If `method == "ridge"` then ridge penalties are only applied to those variables.
If NULL then all variables are penalized. Setting to "batches" will reproduce
original `rcna` behavior.\cr
Default: NULL}

\item{ridges}{A numeric vector of ridge penalties to try. If NULL, a default
set of ridges centered on `svd(X)$d[1]^2` is used.\cr
Default NULL}
}
\value{
A list containing the residuals NAM (NAM_), the annihilator matrix 
used to remove unwanted effects (M), and the approximate degrees of freedom
consumed (r).
}
\description{
remove batch effects (including continuous covariates)
}
\note{
assumes that covariates are not categorical
}
\keyword{internal}
