% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{residNAM}
\alias{residNAM}
\title{remove batch effects (including continuous covariates)}
\usage{
residNAM(
  NAM,
  covs.mat = NULL,
  batch = NULL,
  method = c("ols", "ridge"),
  partial.by = c("batches", "covariates"),
  ridges = NULL,
  verbose = FALSE
)
}
\arguments{
\item{NAM}{a sparseMatrix NAM generated by [buildNAM()]}

\item{covs.mat}{A matrix of covariates.\cr
Default NULL}

\item{batch}{A factor containing batch information.\cr
Default NULL}

\item{method}{A character string specifying what method to use for batch
and covariate correction. Options are: c("ridge", "ols"), Setting to "ridge"
will reproduce original behavior.}

\item{partial.by}{A character string. One of NULL, "batches", or "covariates". If 
`method == "ridge"` then ridge penalties are only applied to those variables.
If null then all variables are penalized. Setting to "batches" will reproduce
original behavior.\cr
Default: NULL}

\item{ridges}{A numeric vector of ridge penalties to try. If NULL, a default
set of ridges is used `c(1e5, 1e4, 1e3, 1e2, 1e1, 1e0, 1e-1, 1e-2, 1e-3, 1e-4, 0)`.\cr
Default NULL}
}
\value{
A list containing the residuals NAM (NAM_), the annihilator matrix used to
remove unwanted effects (M), and teh approximate degrees of freedom (r).
}
\description{
remove batch effects (including continuous covariates)
}
\note{
assumes that covariates are not categorical

the scale function in R gives slightly different results than does 
similar function in python
}
\keyword{internal}
