% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{ridgeNAM}
\alias{ridgeNAM}
\title{Perform batch correction of a NAM using (partial) ridge regression}
\usage{
ridgeNAM(NAM, cov.mat = NULL, batch = NULL, ridges = NULL, partial = NULL)
}
\arguments{
\item{NAM}{A neighborhood association matrix}

\item{cov.mat}{a matrix of covariates. \cr
Default: NULL}

\item{batch}{a factor containing the batch to correct. \cr
Default: NULL}

\item{ridges}{A vector of ridges to try. \cr
Default: NULL}

\item{partial}{A character string or NULL. If NULL, ridge regression is performed
on all variables. If a character, specify which set of variables ridge regression
should be used on, the other set gets OLS fitting. Must be one of "batches" or
"covariates". \cr 
Default: NULL}
}
\value{
A list containing the corrected NAM, the Hat matrix (M), and degrees 
of freedom (r)
}
\description{
Perform batch correction of a NAM using (partial) ridge regression
}
\keyword{internal}
