% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sm.R
\name{Scale}
\alias{Scale}
\title{A scale function for sparseMatrix}
\usage{
Scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A sparseMatrix}

\item{center}{A boolean, should x be centered}

\item{scale}{A boolean, should x be scaled}
}
\value{
A dense sparseMatrix ("dgeMatrix")
}
\description{
A scale function for sparseMatrix
}
\keyword{internal}
