% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nam.R
\name{svdNAM}
\alias{svdNAM}
\title{Compute the singular-value decomposition of a NAM matrix.}
\usage{
svdNAM(NAM, n.pcs = NULL)
}
\arguments{
\item{NAM}{a sparseMatrix NAM generated by [innerNAM]}

\item{n.pcs}{maximum number of components to use in the SVD. Passed to the `k`
parameter of [RSpectra::svds()]. If NULL or `n.pcs > .5 * min(dim(NAM))`
then the full svd is calculated. \cr
Default: NULL}
}
\value{
A list containing the components of the SVD.
}
\description{
Compute the singular-value decomposition of a NAM matrix.
}
\note{
svs are actually eigenvalues, not SVs. I squared them to be consistent 
with python code.

original implementation actually computes dot product: why?

svdNAM scales NAM columns, so this is correlation, not covariance
}
\keyword{internal}
