% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp2.r, R/utils_stats.R
\name{tail_counts}
\alias{tail_counts}
\title{calculate the number of correlations within each threshold}
\usage{
tail_counts(breaks, z)

tail_counts(breaks, z)
}
\arguments{
\item{breaks}{A vector of thresholds}

\item{z}{A matrix of correlations}
}
\value{
a vector of length `ncol(z)` containing the the number of neighborhoods 
exceeding each break (threshold)

a vector of length `ncol(z)` containing the the number of neighborhoods 
exceeding each break (threshold)
}
\description{
calculate the number of correlations within each threshold

calculate the number of correlations within each threshold
}
\keyword{internal}
